style: 디렉토리 구조 수정  - DB, logging 관련 기능은 config 모듈에서 관리하도록
