+++ config/db/connections.py
@@ -1,0 +1,11 @@
+import psycopg2 as pg2
+
+from modules.env import Env
+
+
+class DBConnection:
+    """DB 연결 정보"""
+
+    @staticmethod
+    def get_conn():
+        return pg2.connect(**Env.get_conn_params())
